ownerController 분리
